requirements.txt
alembic==1.5.8
antiorm==1.2.1
click==7.1.2
db==0.1.1
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.4.11
Werkzeug==1.0.1
